refactor(platform): align local transports and file layout
Use UNIX domain sockets for Linux local API clients while retaining named pipes on Windows, with transport-specific coverage. Rename alert, process snapshotter, mock, and NTSTATUS files to match their platform responsibilities.
